align -> assembly and map -> reads
