+++ kaptive/snps.py
@@ -13,9 +13,8 @@ If not, see <https://www.gnu.org/licenses/>.
 import subprocess
 from pathlib import Path
 from tempfile import NamedTemporaryFile
-from typing import List
 
-from .log import warning, quit_with_error
+from kaptive.log import warning, quit_with_error
 
 
 class VcfRecord:
@@ -63,7 +62,7 @@ class VcfRecord:
                f"{self.info}\t{self.format}\t{self.samples}\n"
 
 
-def get_mutation(snps: List[VcfRecord], synonymous: bool = False, all_mutations: bool = False, delim: str = ";") -> str:
+def get_mutation(snps: list[VcfRecord], synonymous: bool = False, all_mutations: bool = False, delim: str = ";") -> str:
     """
     Gets the first disruptive mutation from a list of SNPs generated by `bcftools csq`
     (https://samtools.github.io/bcftools/bcftools.html#csq)
