+++ kaptive/version.py
@@ -14,4 +14,4 @@ details. You should have received a copy of the GNU General Public License along
 If not, see <https://www.gnu.org/licenses/>.
 """
 
-__version__ = '0.0.1'
+__version__ = '3.0.0b'
